+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Changelog
 * Support for new Symfony versions.
 * Support for doctrine/dbal 4.
 * For MySQL/MariaDB, it is now required to configure `defaultTableOptions`.`collate` or `charset` in the Doctrine connection, or alternatively
-  set the encoding explicitly with `Client::setCaseSensitiveEncoding('')` (e.g. `utf8mb4_bin`).
+  set the encoding explicitly with `Client::setCaseSensitiveEncoding()` (e.g. `utf8mb4_bin`).
 * If you are on PHP 8.0 and install Jackalope with `symfony/cache`, you need to restrict `psr/simple-cache` to `^1.0 || ^2.0` in your application because Symfony 5 does not declare a conflict with it, but fails at runtime.
 * Drop support for PHP 7.
 * Fixed: While it is allowed to call `Repository::login` with `null` credentials, there used to be an error. It now correctly works.